Refactor: 컴포넌트 속성(인터페이스) Types 구조화
