+++ src/types/categoryTag.ts
@@ -1,0 +1,5 @@
+export interface CategoryTagProps {
+  label: string;
+  isActive: boolean;
+  onClick?: () => void;
+}
