+++ src/types/heroBanner.ts
@@ -1,0 +1,7 @@
+export interface HeroBannerProps {
+  title?: string;
+  subtitle?: string;
+  discount?: string;
+  backgroundImage?: string | null;
+  onButtonClick?: () => void;
+}
