+++ src/types/navigationBar.ts
@@ -1,0 +1,5 @@
+export interface NavigationBarProps {
+  activeMenu?: string;
+  onSearch?: (value: string) => void;
+  onMenuClick?: (menu: string) => void;
+}
